--- a/src/coreclr/debug/createdump/datatarget.cpp
+++ b/src/coreclr/debug/createdump/datatarget.cpp
@@ -3,6 +3,12 @@
 
 #include "createdump.h"
 
+#if defined(TARGET_ARM64)
+// Flag to check if atomics feature is available on
+// the machine
+bool g_arm64_atomics_present = false;
+#endif
+
 DumpDataTarget::DumpDataTarget(CrashInfo& crashInfo) :
     m_ref(1),
     m_crashInfo(crashInfo)