--- a/src/coreclr/pal/inc/pal.h
+++ b/src/coreclr/pal/inc/pal.h
@@ -74,7 +74,7 @@ extern "C" {
 // On Unix systems, NATIVE_LIBRARY_HANDLE type represents a library handle not registered with the PAL.
 typedef PVOID NATIVE_LIBRARY_HANDLE;
 
-#if defined(TARGET_ARM64)
+#if defined(HOST_ARM64)
 // Flag to check if atomics feature is available on
 // the machine
 extern bool g_arm64_atomics_present;
@@ -3472,7 +3472,7 @@ FORCEINLINE void PAL_ArmInterlockedOperationBarrier()
 #endif
 }
 
-#if defined(TARGET_ARM64)
+#if defined(HOST_ARM64)
 
 #define Define_InterlockMethod(RETURN_TYPE, METHOD_DECL, METHOD_INVOC, INTRINSIC_NAME) \
 __attribute__((target("lse")))  \