--- a/src/coreclr/pal/src/init/pal.cpp
+++ b/src/coreclr/pal/src/init/pal.cpp
@@ -106,6 +106,12 @@ extern "C" BOOL CRTInitStdStreams( void );
 
 extern bool g_running_in_exe;
 
+// #if defined(TARGET_ARM64)
+// // Flag to check if atomics feature is available on
+// // the machine
+// bool g_arm64_atomics_present = false;
+// #endif
+
 Volatile<INT> init_count = 0;
 Volatile<BOOL> shutdown_intent = 0;
 Volatile<LONG> g_coreclrInitialized = 0;