--- a/src/coreclr/pal/src/init/pal.cpp
+++ b/src/coreclr/pal/src/init/pal.cpp
@@ -106,7 +106,7 @@ extern "C" BOOL CRTInitStdStreams( void );
 
 extern bool g_running_in_exe;
 
-#if defined(TARGET_ARM64)
+#if defined(HOST_ARM64)
 // Flag to check if atomics feature is available on
 // the machine
 bool g_arm64_atomics_present = false;