--- a/src/coreclr/utilcode/util.cpp
+++ b/src/coreclr/utilcode/util.cpp
@@ -24,7 +24,7 @@
 #ifndef DACCESS_COMPILE
 UINT32 g_nClrInstanceId = 0;
 
-#if defined(TARGET_WINDOWS) && defined(TARGET_ARM64)
+#if defined(TARGET_ARM64)
 // Flag to check if atomics feature is available on
 // the machine
 bool g_arm64_atomics_present = false;