@kunalspathak 2022-07-04
I tried to incorporate feedback of turning this always ON for OSX/Arm64. Since I don't have that machine, I tried to enable it for Unix/Arm64, just to see it compiles and behave correctly.

Left= Ideal behaviour where we will use `Lse_*` implementation
Right= What it would look like if LSE_DEFAULT_ENABLED=true

![image](https://user-images.githubusercontent.com/12488060/177202519-e0b7a16a-0bdb-4443-acb0-6a57fa3333a9.png)

What I observe is, in Left, we just jump to the inlined implementation where as on right side, it still does some function multiversioning magic and goes to the (I believe) OS implementation of acq_release. See below the point where atomics instructions are executed. 

Here too, left is `LSE_INSTRUCTIONS_ENABLED_BY_DEFAULT=false` and right is `LSE_INSTRUCTIONS_ENABLED_BY_DEFAULT=true`.

![image](https://user-images.githubusercontent.com/12488060/177202939-71f0e264-4de3-47ff-a57f-b4474973530e.png)

I will ask someone to validate what they see on OSX.