@EgorBo 2022-07-14
Improvement son Linux-arm64: https://github.com/dotnet/perf-autofiling-issues/issues/6770